requirements.txt
uvicorn==0.17.0
fastapi==0.72.0
pydantic==1.8.2
SQLAlchemy==1.4.35
alembic==1.7.7
psycopg2-binary==2.9.3
python-dotenv==0.19.1
